+++ src/lidruptracer.hpp
@@ -1,7 +1,8 @@
 #ifndef _lidruptracer_h_INCLUDED
 #define _lidruptracer_h_INCLUDED
 
-class FileTracer;
+#include "tracer.hpp"
+#include "file.hpp"
 
 namespace CaDiCaL {
 
@@ -25,12 +26,12 @@ class LidrupTracer : public FileTracer {
   uint64_t num_clauses;   // number of clauses in hash table
   uint64_t size_clauses;  // size of clause hash table
   LidrupClause **clauses; // hash table of clauses
-  vector<int> imported_clause;
-  vector<int> assumptions;
-  vector<int64_t> imported_chain;
-  vector<int64_t> batch_weaken;
-  vector<int64_t> batch_delete;
-  vector<int64_t> batch_restore;
+  std::vector<int> imported_clause;
+  std::vector<int> assumptions;
+  std::vector<int64_t> imported_chain;
+  std::vector<int64_t> batch_weaken;
+  std::vector<int64_t> batch_delete;
+  std::vector<int64_t> batch_restore;
 
   static const unsigned num_nonces = 4;
 
@@ -60,16 +61,16 @@ class LidrupTracer : public FileTracer {
   void put_binary_lit (int external_lit);
   void put_binary_id (int64_t id, bool = true);
 
-  void lidrup_add_derived_clause (int64_t id, const vector<int> &clause,
-                                  const vector<int64_t> &chain);
-  void lidrup_delete_clause (int64_t id); //, const vector<int> &clause);
+  void lidrup_add_derived_clause (int64_t id, const std::vector<int> &clause,
+                                  const std::vector<int64_t> &chain);
+  void lidrup_delete_clause (int64_t id); //, const std::vector<int> &clause);
   void
-  lidrup_add_restored_clause (int64_t id); //, const vector<int> &clause);
-  void lidrup_add_original_clause (int64_t id, const vector<int> &clause);
-  void lidrup_conclude_and_delete (const vector<int64_t> &conclusion);
+  lidrup_add_restored_clause (int64_t id); //, const std::vector<int> &clause);
+  void lidrup_add_original_clause (int64_t id, const std::vector<int> &clause);
+  void lidrup_conclude_and_delete (const std::vector<int64_t> &conclusion);
   void lidrup_report_status (int status);
-  void lidrup_conclude_sat (const vector<int> &model);
-  void lidrup_conclude_unknown (const vector<int> &trail);
+  void lidrup_conclude_sat (const std::vector<int> &model);
+  void lidrup_conclude_unknown (const std::vector<int> &trail);
   void lidrup_solve_query ();
   void lidrup_batch_weaken_restore_and_delete ();
 
@@ -78,18 +79,18 @@ public:
   ~LidrupTracer ();
 
   // proof section:
-  void add_derived_clause (int64_t, bool, const vector<int> &,
-                           const vector<int64_t> &) override;
-  void add_assumption_clause (int64_t, const vector<int> &,
-                              const vector<int64_t> &) override;
-  void weaken_minus (int64_t, const vector<int> &) override;
-  void delete_clause (int64_t, bool, const vector<int> &) override;
-  void add_original_clause (int64_t, bool, const vector<int> &,
+  void add_derived_clause (int64_t, bool, const std::vector<int> &,
+                           const std::vector<int64_t> &) override;
+  void add_assumption_clause (int64_t, const std::vector<int> &,
+                              const std::vector<int64_t> &) override;
+  void weaken_minus (int64_t, const std::vector<int> &) override;
+  void delete_clause (int64_t, bool, const std::vector<int> &) override;
+  void add_original_clause (int64_t, bool, const std::vector<int> &,
                             bool = false) override;
   void report_status (int, int64_t) override;
-  void conclude_sat (const vector<int> &) override;
-  void conclude_unsat (ConclusionType, const vector<int64_t> &) override;
-  void conclude_unknown (const vector<int> &) override;
+  void conclude_sat (const std::vector<int> &) override;
+  void conclude_unsat (ConclusionType, const std::vector<int64_t> &) override;
+  void conclude_unknown (const std::vector<int> &) override;
 
   void solve_query () override;
   void add_assumption (int) override;
@@ -97,9 +98,9 @@ public:
 
   // skip
   void begin_proof (int64_t) override {}
-  void finalize_clause (int64_t, const vector<int> &) override {}
+  void finalize_clause (int64_t, const std::vector<int> &) override {}
   void strengthen (int64_t) override {}
-  void add_constraint (const vector<int> &) override {}
+  void add_constraint (const std::vector<int> &) override {}
 
   // logging and file io
   void connect_internal (Internal *i) override;