+++ src/tier.cpp
@@ -35,12 +35,18 @@ void Internal::recompute_tier () {
     tier1[stable] = 1;
     tier2[stable] = 1;
     uint64_t accumulated_used = stats.used[stable][0];
-    for (size_t glue = 1; glue < stats.used[stable].size (); ++glue) {
+    size_t glue = 1;
+    for (; glue < stats.used[stable].size (); ++glue) {
       const uint64_t u = stats.used[stable][glue];
       accumulated_used += u;
-      if (accumulated_used <= accumulated_tier1_limit) {
+      if (accumulated_used >= accumulated_tier1_limit) {
         tier1[stable] = glue;
+	break;
       }
+    }
+    for (; glue < stats.used[stable].size (); ++glue) {
+      const uint64_t u = stats.used[stable][glue];
+      accumulated_used += u;
       if (accumulated_used >= accumulated_tier2_limit) {
         tier2[stable] = glue;
         break;
@@ -49,14 +55,9 @@ void Internal::recompute_tier () {
   }
 
   assert (tier1[stable] > 0);
-  if (tier1[stable] >= tier2[stable])
-    tier2[stable] = tier1[stable] + 1;
-  assert (tier2[stable] > tier1[stable]);
 
   assert (tier1[stable]);
   assert (tier2[stable]);
-  if (tier1[stable] == tier2[stable])
-    ++tier2[stable];
 
   if (tier1[stable] < opts.tier1minglue) {
     LOG ("tier1 limit of %d is too low, setting %d instead", tier1[stable], opts.tier1minglue);
@@ -66,6 +67,10 @@ void Internal::recompute_tier () {
     LOG ("tier2 limit of %d is too low, setting %d instead", tier2[stable], opts.tier2minglue);
     tier2[stable] = opts.tier2minglue;
   }
+  if (tier1[stable] >= tier2[stable])
+    tier2[stable] = tier1[stable] + 1;
+  assert (tier2[stable] > tier1[stable]);
+
   PHASE ("retiered", stats.tierecomputed, "tier1 limit = %d in %s mode, tier2 limit = %d in %s mode", tier1[stable],
        stable ? "stable" : "focused", tier2[stable], stable ? "stable" : "focused");
 }
@@ -81,6 +86,10 @@ void Internal::print_tier_usage_statistics () {
     const std::string mode = stable ? "stable" : "focused";
     const size_t tier1 = internal->tier1[stable];
     const size_t tier2 = internal->tier2[stable];
+    if (tier1 > tier2 && opts.reducetier1glue > opts.reducetier2glue) {
+      MSG ("tier1 > tier 2 due to the options, giving up");
+      break;
+    }
     assert (tier1 <= tier2);
     unsigned prefix, suffix;
     unsigned span = tier2 - tier1 + 1;