+++ src/walk.cpp
@@ -43,7 +43,8 @@ struct Walker {
 #ifndef NDEBUG
   std::vector<signed char> current_best_model; // best model found so far
 #endif
-  Walker (Internal *, double size, int64_t limit);
+  Walker (Internal *, int64_t limit);
+  void populate_table (double size);
   void push_flipped (int flipped);
   void save_walker_trail (bool);
   void save_final_minimum (int64_t flips, int64_t old_minimum);
@@ -86,7 +87,7 @@ inline static double fitcbval (double size) {
 
 // Initialize the data structures for one local search round.
 
-Walker::Walker (Internal *i, double size, int64_t l)
+Walker::Walker (Internal *i, int64_t l)
     : internal (i), random (internal->opts.seed), // global random seed
       ticks (0), limit (l), best_trail_pos(-1) {
   random += internal->stats.walk.count; // different seed every time
@@ -95,7 +96,9 @@ Walker::Walker (Internal *i, double size, int64_t l)
 #ifndef NDEBUG
   current_best_model.resize (i->max_var + 1, 0);
 #endif
+}
 
+void Walker::populate_table (double size) {
   // This is the magic constant in ProbSAT (also called 'CB'), which we pick
   // according to the average size every second invocation and otherwise
   // just the default '2.0', which turns into the base '0.5'.
@@ -807,31 +810,9 @@ int Internal::walk_round (int64_t limit, bool prev) {
   PHASE ("walk", stats.walk.count,
          "random walk limit of %" PRId64 " ticks", limit);
 
-  // First compute the average clause size for picking the CB constant.
-  //
-  double size = 0;
-  int64_t n = 0;
-  for (const auto c : clauses) {
-    if (c->garbage)
-      continue;
-    if (c->redundant) {
-      if (!opts.walkredundant)
-        continue;
-      if (!likely_to_be_kept_clause (c))
-        continue;
-    }
-    size += c->size;
-    n++;
-  }
-  double average_size = relative (size, n);
-
-  PHASE ("walk", stats.walk.count,
-         "%" PRId64 " clauses average size %.2f over %d variables", n,
-         average_size, active ());
-
   // Instantiate data structures for this local search round.
   //
-  Walker walker (internal, average_size, limit);
+  Walker walker (internal, limit);
 
   bool failed = false; // Inconsistent assumptions?
 
@@ -895,6 +876,9 @@ int Internal::walk_round (int64_t limit, bool prev) {
 #ifdef LOGGING
     int64_t watched = 0;
 #endif
+
+    double size = 0;
+    int64_t n = 0;
     for (const auto c : clauses) {
 
       if (c->garbage)
@@ -910,6 +894,8 @@ int Internal::walk_round (int64_t limit, bool prev) {
       int satisfied = 0;        // clause satisfied?
 
       int *lits = c->literals;
+      size += c->size;
+      n++;
       const int size = c->size;
 
       // Move to front satisfied literals and determine whether there
@@ -954,6 +940,13 @@ int Internal::walk_round (int64_t limit, bool prev) {
           walker.broken.push_back (c);
       }
     }
+
+    double average_size = relative (size, n);
+    walker.populate_table (average_size);
+    PHASE ("walk", stats.walk.count,
+           "%" PRId64 " clauses average size %.2f over %d variables", n,
+           average_size, active ());
+
 #ifdef LOGGING
     if (!failed) {
       int64_t broken = walker.broken.size ();
@@ -964,6 +957,10 @@ int Internal::walk_round (int64_t limit, bool prev) {
     }
 #endif
   }
+
+  assert (walker.table.size ());
+
+  int64_t old_global_minimum = stats.walk.minimum;
 
   int res; // Tells caller to continue with local search.