+++ test/api/run.sh
@@ -73,7 +73,7 @@ run () {
   then
     src=$tests/$1.c
     language=" -x c"
-    COMPILE="$CXX `echo $CXXFLAGS|sed -e 's,-std=c++11,-std=c11,'`"
+    COMPILE="$CXX `echo $CXXFLAGS|sed -e 's,-std=c++17,-std=c11,'`"
   elif [ -f $tests/$1.cpp ]
   then
     src=$tests/$1.cpp
@@ -87,7 +87,7 @@ run () {
   rm -f $name.log $name.o $name
   status=0
   cmd $COMPILE$language -o $name.o -c $src
-  cmd $COMPILE -o $name $name.o -L$CADICALBUILD -lcadical
+  cmd $COMPILE -o $name $name.o -L$CADICALBUILD $CADICALBUILD/libcadical.a
   cmd $name
   if test $status = 0
   then